Fix conditional dependencies in root packages
SwiftPM's build system always build every target that's part of a root package, regardless of its reachability in the graph, supposedly to ensure all the code in the root package will get exercised. This is of course fundamentally incompatible with the idea of conditional dependencies and therefore causes user confusion.

The most reasonable way to resolve this seems to be to exclude any target that's conditionally depended upon with a condition unfullfiled by the current build environment from building automatically because of its presence in the root package. This does not change what gets built based on dependencies, so degenerate cases where a target is depended upon both conditionally and unconditionally will not be affected. Because of this, it seems reasonable to not tie the new behavior to the tools-version.

rdar://97751068